Fix: WAL partition poll function type (#2301)
* fix: type

* fix: cast to int32

* debug: add logging

* debug: more logs

* Revert "debug: more logs"

This reverts commit 2ff8033f89687e06d07384f6320052160c346e9f.

* Revert "debug: add logging"

This reverts commit a7aaa05b9cd8b027638eae8fbfefae5e19fadd2b.

* fix: rm unnecessary generic

* feat: span attrs + names

* fix: span naming, more details

+++ pkg/repository/v1/sqlcv1/payload-store.sql
@@ -93,7 +93,7 @@ FROM
 WITH tenants AS (
     SELECT UNNEST(
         find_matching_tenants_in_payload_wal_partition(
-            @partitionNumber::BIGINT
+            @partitionNumber::INT
         )
     ) AS tenant_id
 )
